[0.7.1] fixed credit upgrade not costing credits.
changed the diminishing returns formula to a more deterministic formula, at least as much as it can be with as many scaling factors applied to it as there are.
changed the "useShipIdForQualityCalculation" setting to "useRandomQuality" instead, as that's what it was doing effectively, and made it actually random.
changed the ExtraSystemsHM hullmod description to clarify that the quality is not a multiplier, but just a representation of the quality of the ship.
nerfed weapon range and armor bonuses.